requirements-dev.txt
# Dev ops tools
Fabric3==1.14.post1 ; platform_python_implementation != 'PyPy'

# Paramiko deps (used by Fabric)
cryptography==2.7
pyasn1==0.4.8
PyNaCl==1.4.0

paramiko==2.7.1

# Testing
pytest>=5.2.0
pytest_selenium>=1.17.0
selenium==3.14.0
pytest_bdd>=3.2.1
pytest-flask>=0.15.0
pytest-xdist>=1.29.0
pytest-metadata>=1.8.0
junit-xml>=1.9
deepdiff==5.5.0
ordered-set==4.0.2

# Other dev deps
pip-licenses==1.7.1
termcolor==1.1.0
googlemaps==4.2.0
geopy==2.1.0

# Set pylint and asteroid (the code parser used by pylint) to these versions
# because later versions have significant performance issues when working with
# code that imports pandas.
# https://github.com/PyCQA/pylint/issues/2198
pylint>=2.3.1
black==21.5b2

# Pyinstaller is what we use to build binaries
PyInstaller>=3.4

pytest-env==0.6.2
coverage==5.2.1
pytest-picked==0.4.4

# mypy and related typing dependencies
mypy==0.910
sqlalchemy-stubs==0.4
types-certifi==0.1.3
types-Flask==0.1.2
types-Jinja2==0.1.3
types-jwt==0.1.3
types-python-dateutil==0.1.2
types-redis==0.1.17
types-requests==0.1.8
types-Werkzeug==0.1.8

# We use python-jenkins to trigger jenkins builds remotely
python-jenkins==1.7.0

testfixtures==6.17.1
alchemy-mock==0.4.3
factory-boy==3.2.0
